Update Asynchronous-Discussions.md
I took out "Flipgrid has a $60 annual fee for staff to create and use grids but is free for students to access and use." I think it should be assumed (as per the CSU policy/guidelines) that any external tools are free for students. Cost for staff should not be on this site I think.